+++ src/util/Vector2.java
@@ -1,15 +1,15 @@
 package util;
 
 public class Vector2 {
-	public static int X, Y;
+	public static Double X, Y;
 	
-	public Vector2(int X, int Y){
+	public Vector2(Double X, Double Y){
 		Vector2.X = X;
 		Vector2.Y = Y;
 	}
 	
 	public Vector2(){
-		Vector2.X = 0;
-		Vector2.Y = 0;
+		Vector2.X = 0.0;
+		Vector2.Y = 0.0;
 	}
 }
