+++ src/util/Vector3.java
@@ -1,16 +1,16 @@
 package util;
 
 public class Vector3 {
-	public static int X, Y, Z;
-	public Vector3(int X, int Y, int Z){
+	public static Double X, Y, Z;
+	public Vector3(Double X, Double Y, Double Z){
 		Vector3.X = X;
 		Vector3.Y = Y;
 		Vector3.Z = Z;
 	}
 	
 	public Vector3(){
-		Vector3.X = 0;
-		Vector3.Y = 0;
-		Vector3.Z = 0;
+		Vector3.X = 0.0;
+		Vector3.Y = 0.0;
+		Vector3.Z = 0.0;
 	}
 }
